logica crear tarjetas

+++ HomeBankingMindHub/Utils/RandomNumber.cs
@@ -5,10 +5,10 @@ namespace HomeBankingMindHub.Utils
 {
     public static class NumberGenerator
     {
-        public static int GenerarNumero(int lowerBound, int upperBound)
+        public static int GenerarNumero(long lowerBound, long upperBound)
         {
                 Random random = new Random();
-                int randomNumber = random.Next(lowerBound, upperBound);
+                int randomNumber = random.Next((int)lowerBound, (int)upperBound);
                 return randomNumber;
         }
     }
